refactor(user-info.service): deleteUserInfo 수정
UserInfo가 존재하는지 확인하는 로직 추가

+++ personal-blog-service/src/user/service/user-info.service.ts
@@ -55,6 +55,11 @@ export class UserInfoService {
   }
 
   async deleteUserInfoByUid(uid: string): Promise<void> {
+    const isExist = await this.userInfoRepository.isExist(uid);
+    if (!isExist) {
+      throw new NotFoundException(`UserInfo does not exist. - [${uid}]`);
+    }
+
     await this.userInfoRepository.deleteUserInfoByUid(uid);
 
     this.logger.info(`UserInfo has been deleted. - [${uid}]`);
